field and tree optimised

+++ src/main/java/org/banta/citronix/dto/field/UpdateFieldRequest.java
@@ -7,5 +7,5 @@ import java.util.UUID;
 @Data
 public class UpdateFieldRequest {
     @Min(value = 1000, message = "Field area must be at least 0.1 hectare (1000 m²)")
-    private Float area;
+    private Double area;
 }
